requirements.txt
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
scipy==1.10.1
torch==2.0.1
torchvision==0.15.2